Move readable_tracks property to File classes
